@jchen10 2025-06-23
The perf improvement on LNL: 548->705, 28%.

Without the PR,
```
model_benchmark.exe -i ..\models\phi3.5-web-accuracy4-gqa --prompt_file prompt.txt -g 128 -r 1
Batch size: 1, prompt tokens: 1024, tokens to generate: 128
Prompt processing (time to first token):
        avg (us):       1.86824e+06
        avg (tokens/s): 548.108
        p50 (us):       1.86824e+06
        stddev (us):    0
        n:              1 * 1024 token(s)
Token generation:
        avg (us):       35224.9
        avg (tokens/s): 28.389
        p50 (us):       35016.8
        stddev (us):    2027.73
        n:              127 * 1 token(s)
Token sampling:
        avg (us):       7
        avg (tokens/s): 142857
        p50 (us):       7
        stddev (us):    0
        n:              1 * 1 token(s)
E2E generation (entire generation loop):
        avg (ms):       6341.9
        p50 (ms):       6341.9
        stddev (ms):    0
        n:              1
Peak working set size (bytes): 4398538752
```

With the PR,
```
model_benchmark.exe -i ..\models\phi3.5-web-accuracy4-gqa --prompt_file prompt.txt -g 128 -r 1
Batch size: 1, prompt tokens: 1024, tokens to generate: 128
Prompt processing (time to first token):
        avg (us):       1.45052e+06
        avg (tokens/s): 705.952
        p50 (us):       1.45052e+06
        stddev (us):    0
        n:              1 * 1024 token(s)
Token generation:
        avg (us):       35316.4
        avg (tokens/s): 28.3154
        p50 (us):       34994
        stddev (us):    2197.23
        n:              127 * 1 token(s)
Token sampling:
        avg (us):       7.2
        avg (tokens/s): 138889
        p50 (us):       7.2
        stddev (us):    0
        n:              1 * 1 token(s)
E2E generation (entire generation loop):
        avg (ms):       5935.81
        p50 (ms):       5935.81
        stddev (ms):    0
        n:              1
Peak working set size (bytes): 4398870528
```