@jchen10 2025-06-24
> ## Pull Request Overview
> This PR optimizes the Intel path for `subgroup_matrix_matmul_nbits` by removing the per-thread shared-memory load of matrix A and instead using `subgroupMatrixLoad` directly from global memory, reducing SLM usage and bandwidth pressure.
> 
> * Removed `var<workgroup> tile_A` and the `loadSHMA` helper function.
> * Updated inner loop to compute a global offset and call `subgroupMatrixLoad` on `input_a`.
> * Adjusted indexing and stride parameters to match the global layout.
> 
> Comments suppressed due to low confidence (2)

Copilot AI review looks really cool. Very good summary, even better than what I wrote.